// src/app/model/Agent.ts
export class Agent {


    agentId: number = 0;
    agentFirstName: string = '';
    agentLastName: string = '';
    qualification: string = '';
    email: string = '';
    phone: number = 0;
    username: string = '';
    password: string = '';
    status: boolean = true;
    customerCount: number = 0;
    userId: number = 0;
}